backend/requirements.txt
azure-functions>=1.19.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.0.0
openai>=1.0.0
python-dateutil>=2.8.2
